Add ContextInputs type
